requirements.txt
addict==2.4.0
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.0
cycler==0.12.1
dash==3.2.0
decorator==5.2.1
drjit==1.2.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.4
imath==0.0.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.2.1
mitsuba==3.7.1
mpmath==1.3.0
narwhals==2.10.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
objprint==0.3.0
open3d==0.19.0
opencv-python==4.9.0.80
OpenEXR==3.4.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
plyfile==1.1.3
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pymeshlab==2025.7
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.10.23
requests==2.28.1
retrying==1.4.2
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.15.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.35.2
trimesh==4.9.0
triton==2.1.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.13
varname==0.15.0
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zipp==3.23.0
